+++ NEWS.md
@@ -2,7 +2,7 @@
 
 ## New Features
 
-- You can link the GitHub Issue to Changelog by using (#ISSUE-NUMBER)
+- Added a function to derive the variable STUDYEYE in ADSL (issue #9).
 
 ## Updates of Existing Functions