+++ NEWS.md
@@ -3,6 +3,7 @@
 ## New Features
 
 - Added a function to derive the variable STUDYEYE in ADSL (issue #9).
+- Added a function to derive the variable AFEYE in Occurrence datasets (issue #10).
 
 ## Updates of Existing Functions