+++ NEWS.md
@@ -1,9 +1,10 @@
-# admiralophtha 0.0.1
+# admiralophtha 0.1.0
 
 ## New Features
 
 - Added a function to derive the variable STUDYEYE in ADSL (issue #9).
 - Added a function to derive the variable AFEYE in Occurrence datasets (issue #10).
+- Added a function to derive the criterion flags in ADBCVA (issue #49).
 
 ## Updates of Existing Functions